refactor page; refactor roadmap items into markdown files
