+++ package.json
@@ -59,6 +59,7 @@
     "@tailwindcss/line-clamp": "^0.4.0",
     "@tailwindcss/typography": "^0.5.3",
     "@tsconfig/docusaurus": "^1.0.4",
+    "gray-matter": "^4.0.3",
     "remark-code-import": "^0.4.0",
     "typescript": "^4.6.2",
     "webpack": "^5.72.0",
