+++ roadmap/1_core-protocol/deployed/network_performance.md
@@ -1,0 +1,9 @@
+---
+title: "Network Performance with Larger Network: State Sync, Certification, and XNet"
+links:
+  Proposal: https://dashboard.internetcomputer.org/proposal/35648
+  Forum Post: https://forum.dfinity.org/t/long-term-r-d-scalability-proposal/9387/3
+eta: February 2022
+---
+
+This feature ensures that the Internet Computer meets future scalability requirements in terms of number of subnets and size of their growing canister state. The main focus is on the scalability of the XNet communication protocol and the state sync protocol, including state certification.
