+++ roadmap/1_core-protocol/in-progress/btc-integration.md
@@ -1,0 +1,10 @@
+---
+title: Bitcoin Integration to Enable Bitcoin Smart Contracts
+links:
+  Forum Post: https://forum.dfinity.org/t/direct-integration-with-bitcoin/6147
+  Proposal: https://dashboard.internetcomputer.org/proposal/20586
+eta: August 2022
+is_community: false
+---
+
+Bitcoin, the world's first blockchain, has evolved to the world's digital store of value and is therefore often referred to as a digital version of gold. We directly integrate the Internet Computer with the Bitcoin blockchain, i.e., canisters can themselves hold and transfer bitcoin. Importantly, direct integration means that no additional trust assumptions are required and indeed no additional parties, such as bridges, are needed. This feature relies on threshold ECDSA signatures that make it possible for a subnet to sign on behalf of a canister with a secret-shared key. This feature will enable for the first time smart contracts for Bitcoin leveraging the countless powerful features of canisters.
