+++ roadmap/1_core-protocol/in-progress/deterministic-time-slicing.md
@@ -1,0 +1,10 @@
+---
+title: Deterministic Time Slicing
+links:
+  Forum Post: https://forum.dfinity.org/t/deterministic-time-slicing/10635
+  Proposal:
+eta:
+is_community: false
+---
+
+Currently, the computation a canister call can perform is hard bounded per round. Deterministic time slicing allows for long(er) running, multi-round computations by suspending the execution at the end of one round and resuming it in the next.
