+++ roadmap/1_core-protocol/in-progress/high-replication-subnets.md
@@ -1,0 +1,8 @@
+---
+title: High-replication subnets
+links:
+  Forum Post:
+  Proposal:
+eta:
+is_community: false
+---
