+++ roadmap/1_core-protocol/in-progress/http-requests.md
@@ -1,0 +1,10 @@
+---
+title: Enable HTTP Requests from Canisters
+links:
+  Forum Post: https://forum.dfinity.org/t/long-term-r-d-general-integration-proposal/9383
+  Proposal: https://dashboard.internetcomputer.org/proposal/35639
+eta:
+is_community: false
+---
+
+This feature enables canister (dapps/smart contracts) on the Internet Computer to make HTTP(S) requests to services outside the IC and thus integrates the Internet Computer with the Web 2.0 world. This enables a plurality of new use cases, e.g., directly obtaining exchange rate data from external servers for DeFi dapps, obtaining weather data for decentralized insurance services, or sending notifications to end users via traditional communications channels. The first version of this feature will cover a subset of the typical oracle service functionalities, but will do so in a trustless manner, i.e., without making any additional trust assumptions.
