+++ roadmap/1_core-protocol/in-progress/subnet-rental.md
@@ -1,0 +1,8 @@
+---
+title: Subnet rental
+links:
+  Forum Post:
+  Proposal:
+eta:
+is_community: false
+---
