+++ roadmap/1_core-protocol/in-progress/tecdsa.md
@@ -1,0 +1,10 @@
+---
+title: Threshold ECDSA signatures
+links:
+  Forum Post: https://forum.dfinity.org/t/threshold-ecdsa-signatures/6152
+  Proposal: https://dashboard.internetcomputer.org/proposal/21340
+eta: August 2022
+is_community: false
+---
+
+ECDSA signatures are widely used in the blockchain industry. This feature will enable canister smart contracts to have an ECDSA public key and to sign with regard to it. The corresponding secret key is threshold-shared among the nodes of the subnet holding the canister smart contract. This is a prerequisite for the direct integration between the Internet Computer and Bitcoin and Ethereum.
