+++ roadmap/1_core-protocol/index.md
@@ -1,0 +1,8 @@
+---
+title: Core Protocol
+
+card: /img/roadmap/core-protocol.card.jpg
+overlay: /img/roadmap/core-protocol.overlay.jpg
+---
+
+Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
