+++ roadmap/1_core-protocol/pending/eth-integration.md
@@ -1,0 +1,14 @@
+---
+title: Direct Integration with Ethereum
+links:
+  Forum Post: https://forum.dfinity.org/t/long-term-r-d-integration-with-the-ethereum-network/9382/6
+  Proposal: https://dashboard.internetcomputer.org/proposal/35635
+eta:
+is_community: false
+---
+
+The Direct Integration with Ethereum and the Internet Computer plans to encompass a few implementations, as follows:
+
+Ethereum blockchain integration: Integrating the Internet Computer blockchain with the Ethereum blockchain in a trustless manner, i.e., without using any trusted intermediaries such as bridges. This will enable smart contracts on the Internet Computer to call smart contracts on Ethereum and vice versa.
+
+EVM support on the Internet Computer (IC): Providing an Ethereum execution environment (EVM) on the Internet Computer to run Solidity-/EVM-based smart contracts on the IC.
