+++ roadmap/2_boundary_nodes/deployed/icos-nodes.md
@@ -1,0 +1,10 @@
+---
+title: ICOS Boundary Nodes
+links:
+  Forum Post: https://forum.dfinity.org/t/long-term-r-d-boundary-nodes-proposal/9401
+  Proposal: https://dashboard.internetcomputer.org/proposal/35671
+eta: August 2022
+is_community: false
+---
+
+Today, the installation and configuration of boundary nodes requires manual intervention by the foundation’s employees. This feature establishes a fully-automated build process to facilitate improved automated testing and deployment and thereby enables NNS-driven deployment of boundary nodes. The operating system, IC OS, currently used for replica nodes, shall also be used for boundary nodes. This solution builds the basis for ongoing decentralization efforts.
