+++ roadmap/2_boundary_nodes/in-progress/seo_and_streaming.md
@@ -1,0 +1,10 @@
+---
+title: SEO and Streaming Support
+links:
+  Forum Post:
+  Proposal:
+eta: Sept 2022
+is_community: false
+---
+
+This feature enables users on the IC to have their dapps indexed by search engines and to provide previews on social media (e.g., Twitter cards). To this end, boundary nodes redirect requests from crawlers and bots (e.g., Googlebot) to raw in order to avoid loading the service worker.
