+++ roadmap/2_boundary_nodes/pending/custom_subdomains.md
@@ -1,0 +1,10 @@
+---
+title: Custom Subdomains
+links:
+  Forum Post: https://forum.dfinity.org/t/custom-domains-for-ic0-app-community-consideration/6162
+  Proposal:
+eta:
+is_community: false
+---
+
+CanisterId-based URLs, such as ​​https://7e6iv-biaaa-aaaaf-aaada-cai.ic0.app, are hard to read and memorize. This feature will enable developers on the platform to create custom subdomains on ic0.app for their canisters. For instance, a developer could create a custom canister name <name> for their canister such that <name>.ic0.app resolves to their <canister_id>. This is the first step towards enabling arbitrary URLs.
