+++ roadmap/6_infrastructure/index.md
@@ -1,0 +1,7 @@
+---
+title: Infrastructure & Operations
+card: /img/roadmap/infrastructure.card.jpg
+overlay: /img/roadmap/infrastructure.overlay.jpg
+---
+
+Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
